requirements - bot.txt
numpy
onnxruntime
sounddevice
soundfile
discord.py
gdown
praat-parselmouth
markdown